+++ samples/sign-message/src/config.ts
@@ -2,6 +2,9 @@ import {
     BrowserWalletConnector,
     CONCORDIUM_WALLET_CONNECT_PROJECT_ID,
     WalletConnectConnector,
+    WalletConnectEvent,
+    WalletConnectMethod,
+    WalletConnectNamespaceConfig,
     ephemeralConnectorType,
 } from '@concordium/react-components';
 import { SignClientTypes } from '@walletconnect/types';
@@ -16,5 +19,12 @@ const WALLET_CONNECT_OPTS: SignClientTypes.Options = {
     },
 };
 
+const WALLET_CONNECT_SCOPE: WalletConnectNamespaceConfig = {
+    methods: [WalletConnectMethod.SignMessage],
+    events: [WalletConnectEvent.AccountsChanged, WalletConnectEvent.ChainChanged],
+};
+
 export const BROWSER_WALLET = ephemeralConnectorType(BrowserWalletConnector.create);
-export const WALLET_CONNECT = ephemeralConnectorType(WalletConnectConnector.create.bind(this, WALLET_CONNECT_OPTS));
+export const WALLET_CONNECT = ephemeralConnectorType((delegate, network) =>
+    WalletConnectConnector.create(WALLET_CONNECT_OPTS, delegate, network, WALLET_CONNECT_SCOPE)
+);