+++ packages/wallet-connectors/src/WalletConnect.ts
@@ -432,6 +432,12 @@ export class WalletConnectConnection implements WalletConnection {
     }
 
     async signMessage(accountAddress: string, msg: SignableMessage) {
+        const connectedAccount = this.getConnectedAccount();
+        if (accountAddress !== connectedAccount) {
+            throw new Error(
+                `cannot sign message with address '${accountAddress}' on connection for account '${connectedAccount}'`
+            );
+        }
         const params = convertSignableMessageFormat(msg);
         // TODO Catch thrown non-Error and rethrow it as a proper Error.
         const signature = await this.connector.client.request({